requirements.txt
django==3.0.3
djangorestframework==3.11.0
markdown==3.1.1
django-filter==2.2.0
django-environ==0.4.5
drf-yasg==1.17.0
psycopg2-binary==2.8.4
python-telegram-bot==12.3.0
django-rest-auth==0.9.5
django-allauth==0.41.0
